Made physical attributes into decimal values (floats)

+++ Engine/Dude.h
@@ -9,13 +9,13 @@ public:
 	void ClampToScreen();
 	void Draw( Graphics& gfx ) const;
 	void Update( const Keyboard& kbd );
-	int GetX() const;
-	int GetY() const;
-	int GetWidth() const;
-	int GetHeight() const;
+	float GetX() const;
+	float GetY() const;
+	float GetWidth() const;
+	float GetHeight() const;
 private:
-	int x = 400;
-	int y = 300;
+	float x = 400.0f;
+	float y = 300.0f;
 	static constexpr int speed = 1;
 	static constexpr int width = 20;
 	static constexpr int height = 20;
